One shared cfg_exclusive, in crate::cfg (#357 review)
The predicate-exclusivity test moves from the PyO3 scan to `crate::cfg`, next
to `predicate_string` whose rendering it reads, and the PyO3 collision
markers delegate to it. No behaviour change: the `#[julia]` side keeps its
broader, deliberate exemption — an undecided `#[cfg]` never claims, which is
what lets `#[cfg(unix)]` and `#[cfg(windows)]` variants of one item coexist
though neither is syntactically the other's negation
(`symbol_owners_feed_the_duplicate_check`) — and the PyO3 side keeps the
narrower "only when provably exclusive" one, which is strictly better than the
no-exemption it had before this branch. Full suite green.
